Move types out into their own place
